requirements.txt
google-genai
PyPDF2
python-docx
tika
faiss-cpu
numpy
